master merged

+++ src/main/java/com/muates/springbootbookstore/service/BookService.java
@@ -1,6 +1,8 @@
 package com.muates.springbootbookstore.service;
 
+import com.muates.springbootbookstore.domain.Author;
 import com.muates.springbootbookstore.domain.Book;
+import com.muates.springbootbookstore.domain.Publisher;
 import com.muates.springbootbookstore.repository.BookRepository;
 import org.springframework.stereotype.Service;
 
@@ -14,9 +16,13 @@ import java.util.Optional;
 public class BookService {
 
     private final BookRepository bookRepository;
+    private final PublisherService publisherService;
+    private final AuthorService authorService;
 
-    public BookService(BookRepository bookRepository) {
+    public BookService(BookRepository bookRepository, PublisherService publisherService, AuthorService authorService) {
         this.bookRepository = bookRepository;
+        this.publisherService = publisherService;
+        this.authorService = authorService;
     }
 
     public List<Book> getAllBooks(){
@@ -27,8 +33,12 @@ public class BookService {
         return bookRepository.findById(id).orElseThrow(() -> new NoSuchElementException("User does not found!"));
     }
 
-    public void saveBook(Book book){
-        bookRepository.save(book);
+    public Book saveBook(Book book){
+        Publisher publisher = publisherService.getPublisherById(book.getPublisher().getId());
+        Author author = authorService.getAuthorById(book.getAuthor().getId());
+        book.setPublisher(publisher);
+        book.setAuthor(author);
+        return bookRepository.save(book);
     }
 
     public void updateBookById(Long id, Book book){
@@ -46,6 +56,9 @@ public class BookService {
     }
 
     public void deleteBookById(Long id){
-        bookRepository.deleteById(id);
+        Book book = getBookById(id);
+        if(book != null){
+            bookRepository.deleteById(id);
+        }
     }
 }
